division reele

+++ Aquisition_axi.c
@@ -102,7 +102,7 @@ int main(int argc, char **argv)
     fprintf(fichier, "\n");
 
     for (int i = 0; i < dsize; i++) {
-        float time = (i/SAMPLE_RATE)*dec;
+        float time = (float)(i/SAMPLE_RATE)*dec;
         fprintf(fichier, "%f", time);
         if (i!= dsize -1) fprintf(fichier, ",");
     }
